Added the function to move to slide
Buttons are now created with js, one for each image.
Both images and buttons share the same
id number (img-x, btn-x).
Added the function to move to a particular slide by
clicking on the "skip" buttons. If clicking the same
does nothing.
Need to make sure the order of the images stays the same
otherwise, when mixing next/previous buttons and skip
buttons clicks we don't want users to go looking for the
image they saw but can't find.
Need to color the button selected so that we know at what
image we are.
Need to make the next/previous functions.