Added goNext function
We can use the forward button now.

Need to do the same for backward button but should be easy.
Need the 5sec timeout.
Buttons and images are changed based on id so the second
button will always display the second image and so on.
Next and previous functions also make use of the id to change
between images. What might be a problem is in the case of a
carousel that shows all the images, their order might not be
the same as that of the id order because we are appending
images to the div so their order is messed up if you skip
from one image to another. Therefore, we need to sort images
based on id after we swap them. This is not a problem if the
carousel doesn't show previous and next images but only the
current image.